code refactoring

+++ src/com/sleep/component/death/DeathComponent.java
@@ -3,22 +3,40 @@ package com.sleep.component.death;
 import com.sleep.GameScreen;
 import com.sleep.component.Component;
 import com.sleep.component.ComponentException;
+import com.sleep.component.movement.MovementComponent;
 
+/**
+ * when die() is used, the entity is instantly removed from the grid.
+ * 
+ * the entity is only removed from the entityList though, when certain critera
+ * have been met.
+ * 
+ */
 public class DeathComponent extends Component {
-	
+	private boolean alive = true;
+
 	public void die() {
-		// oh no
-		GameScreen.entityManager.remove(owner);
+		alive = false;
+		GameScreen.grid.removeEntity(owner);
 	}
 
 	@Override
 	public void update(float delta) {
-		
+
+		if (!alive) {
+			MovementComponent moveComp = owner.getComponent(MovementComponent.class);
+			if (moveComp != null) {
+				if (moveComp.isMoving()) {
+					return;
+				}
+			}
+			GameScreen.entityManager.remove(owner);
+		}
 	}
 
 	@Override
 	public void init() throws ComponentException {
-		
+
 	}
 
 }
